wip: extract the coordinates translation code
