+++ src/component/mxgraph/MxGraphRenderer.ts
@@ -22,8 +22,10 @@ import { MxGraphFactoryService } from '../../service/MxGraphFactoryService';
 import Waypoint from '../../model/bpmn/edge/Waypoint';
 import { StyleConstant } from './StyleUtils';
 import { Font } from '../../model/bpmn/Label';
+import CoordinatesTranslator from './extension/CoordinatesTranslator';
 
-interface Coordinate {
+// TODO move
+export interface Coordinate {
   x: number;
   y: number;
 }
@@ -31,7 +33,13 @@ interface Coordinate {
 export default class MxGraphRenderer {
   private mxPoint: typeof mxgraph.mxPoint = MxGraphFactoryService.getMxGraphProperty('mxPoint');
   private mxConstants: typeof mxgraph.mxConstants = MxGraphFactoryService.getMxGraphProperty('mxConstants');
-  constructor(readonly graph: mxgraph.mxGraph) {}
+
+  private readonly coordinatesTranslator: CoordinatesTranslator;
+
+  constructor(readonly graph: mxgraph.mxGraph) {
+    // TODO temp prior injection
+    this.coordinatesTranslator = new CoordinatesTranslator(graph);
+  }
 
   public render(bpmnModel: BpmnModel): void {
     const model = this.graph.getModel();
@@ -150,31 +158,32 @@ export default class MxGraphRenderer {
   }
 
   private getRelativeCoordinates(parent: mxgraph.mxCell, absoluteCoordinate: Coordinate): Coordinate {
-    const translateForRoot = this.getTranslateForRoot(parent);
-    const relativeX = absoluteCoordinate.x + translateForRoot.x;
-    const relativeY = absoluteCoordinate.y + translateForRoot.y;
-    return { x: relativeX, y: relativeY };
+    return this.coordinatesTranslator.computeRelativeCoordinates(parent, absoluteCoordinate);
+    // const translateForRoot = this.getTranslateForRoot(parent);
+    // const relativeX = absoluteCoordinate.x + translateForRoot.x;
+    // const relativeY = absoluteCoordinate.y + translateForRoot.y;
+    // return { x: relativeX, y: relativeY };
   }
 
-  // Returns the translation to be applied to a cell whose mxGeometry x and y values are expressed with absolute coordinates
-  // (i.e related to the graph default parent) you want to assign as parent to the cell passed as argument of this function.
-  // That way, you will be able to express the cell coordinates as relative to its parent cell.
+  // // Returns the translation to be applied to a cell whose mxGeometry x and y values are expressed with absolute coordinates
+  // // (i.e related to the graph default parent) you want to assign as parent to the cell passed as argument of this function.
+  // // That way, you will be able to express the cell coordinates as relative to its parent cell.
+  // //
+  // // The implementation taken from the example described in the documentation of mxgraph#getTranslateForRoot
+  // // The translation is generally negative
+  // private getTranslateForRoot(cell: mxgraph.mxCell): mxgraph.mxPoint {
+  //   const model = this.graph.getModel();
+  //   const offset = new this.mxPoint(0, 0);
   //
-  // The implementation taken from the example described in the documentation of mxgraph#getTranslateForRoot
-  // The translation is generally negative
-  private getTranslateForRoot(cell: mxgraph.mxCell): mxgraph.mxPoint {
-    const model = this.graph.getModel();
-    const offset = new this.mxPoint(0, 0);
-
-    while (cell != null) {
-      const geo = model.getGeometry(cell);
-      if (geo != null) {
-        offset.x -= geo.x;
-        offset.y -= geo.y;
-      }
-      cell = model.getParent(cell);
-    }
-
-    return offset;
-  }
+  //   while (cell != null) {
+  //     const geo = model.getGeometry(cell);
+  //     if (geo != null) {
+  //       offset.x -= geo.x;
+  //       offset.y -= geo.y;
+  //     }
+  //     cell = model.getParent(cell);
+  //   }
+  //
+  //   return offset;
+  // }
 }
